Refactored: Datadog::Quantization --> Datadog::Core::Quantization
