+++ lib/ddtrace.rb
@@ -9,8 +9,8 @@ require 'ddtrace/version'
 require 'datadog/tracing'
 require 'ddtrace/pin'
 require 'ddtrace/tracer'
-require 'ddtrace/quantization/hash'
-require 'ddtrace/quantization/http'
+require 'datadog/core/quantization/hash'
+require 'datadog/core/quantization/http'
 require 'ddtrace/pipeline'
 require 'ddtrace/configuration'
 require 'ddtrace/metrics'
