+++ lib/ddtrace/contrib/mongodb/parsers.rb
@@ -1,3 +1,5 @@
+require 'datadog/core/quantization/hash'
+
 # typed: true
 module Datadog
   module Contrib
@@ -16,21 +18,24 @@ module Datadog
       # returns a formatted and normalized query
       def query_builder(command_name, database_name, command)
         # always exclude the command name
-        options = Quantization::Hash.merge_options(quantization_options, exclude: [command_name.to_s])
+        options = Core::Quantization::Hash.merge_options(quantization_options, exclude: [command_name.to_s])
 
         # quantized statements keys are strings to avoid leaking Symbols in older Rubies
         # as Symbols are not GC'ed in Rubies prior to 2.2
-        base_info = Quantization::Hash.format({
-                                                'operation' => command_name,
-                                                'database' => database_name,
-                                                'collection' => command.values.first
-                                              }, options)
+        base_info = Core::Quantization::Hash.format(
+          {
+            'operation' => command_name,
+            'database' => database_name,
+            'collection' => command.values.first
+          },
+          options
+        )
 
-        base_info.merge(Quantization::Hash.format(command, options))
+        base_info.merge(Core::Quantization::Hash.format(command, options))
       end
 
       def quantization_options
-        Datadog::Quantization::Hash.merge_options(DEFAULT_OPTIONS, configuration[:quantize])
+        Core::Quantization::Hash.merge_options(DEFAULT_OPTIONS, configuration[:quantize])
       end
 
       def configuration
